+++ src-ui/app/components/event-log-entry/entry-parsers/sleep-mode-disabled.ts
@@ -17,6 +17,7 @@ export class EventLogSleepModeDisabledEntryParser extends EventLogEntryParser<Ev
       case 'AUTOMATION':
         switch (entry.reason.automation) {
           case 'SLEEP_MODE_CHANGE_ON_STEAMVR_STATUS':
+          case 'SLEEP_MODE_DISABLE_ON_UPRIGHT_POSE':
           case 'SLEEP_MODE_DISABLE_AT_TIME':
           case 'SLEEP_MODE_DISABLE_AFTER_TIME':
           case 'SLEEP_MODE_DISABLE_ON_DEVICE_POWER_ON':