+++ src-ui/app/services/sleep-detection-automations/sleep-mode-disable-after-time-automation.service.ts
@@ -20,11 +20,7 @@ export class SleepModeDisableAfterTimeAutomationService {
   sleepEnabled = false;
   threshold = -1;
 
-  constructor(
-    private automationConfig: AutomationConfigService,
-    private openvr: OpenVRService,
-    private sleep: SleepService
-  ) {}
+  constructor(private automationConfig: AutomationConfigService, private sleep: SleepService) {}
 
   async init() {
     this.automationConfig.configs